Green do_not_delete_category_when_id_is_zero

+++ src/Domain/UseCases/Static/Category/DeleteCategory.cs
@@ -7,6 +7,9 @@ public class DeleteCategory {
     }
 
     public async Task Execute(int id) {
+        if (id == 0) {
+            throw new ArgumentException("Category code is required");
+        }
         await staticSynchronizerApiClient.DeleteCategory(id.ToString());
     }
 }
